# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Wikitruth API",
    "version": "1.1.0",
    "description": "Stable v1 contract for Wikitruth browser, civic-tenant, administrative, and accountable software-agent integrations."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Stable v1"
    },
    {
      "url": "/api",
      "description": "Compatibility alias"
    }
  ],
  "paths": {
...
    "/civic/platform/tenants": {
      "get": {
        "summary": "List civic tenants (platform administrator)",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tenants": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CivicTenant"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Platform administrator required"
          }
        }
      },
      "post": {
        "summary": "Create a civic tenant (platform administrator)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CivicTenant"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Invalid tenant configuration"
          },
          "403": {
            "description": "Platform administrator required"
          }
        }
      }
    },
    "/civic/platform/tenants/{managedTenantId}": {
      "put": {
        "summary": "Update a civic tenant (platform administrator)",
        "parameters": [
          {
            "name": "managedTenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CivicTenant"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated"
          },
          "404": {
            "description": "Tenant not found"
          }
        }
      }
    },
    "/civic/platform/tenants/bootstrap": {
      "post": {
        "summary": "Persist built-in civic tenant definitions",
        "responses": {
          "200": {
            "description": "Bootstrapped"
          },
          "403": {
            "description": "Platform administrator required"
          }
        }
      }
    },
    "/civic/admin/jurisdictions": {
      "post": {
        "summary": "Create a tenant jurisdiction",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CivicJurisdictionMutation"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Invalid jurisdiction"
          },
          "403": {
            "description": "Tenant administrator required"
          }
        }
      }
    },
    "/civic/admin/memberships": {
      "get": {
        "summary": "List tenant memberships",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "memberships": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CivicTenantMembership"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Tenant administrator required"
          }
        }
      }
    },
    "/civic/admin/memberships/{userId}": {
      "put": {
        "summary": "Create or update tenant membership",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated"
          },
          "403": {
            "description": "Tenant administrator required"
          }
        }
      }
    },
...
    "/tenants/{tenantId}/civic/admin/jurisdictions": {
      "post": {
        "summary": "Create a tenant jurisdiction",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CivicJurisdictionMutation"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Invalid jurisdiction"
          },
          "403": {
            "description": "Tenant administrator required"
          }
        },
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9][a-z0-9-]{1,62}$"
            }
          }
        ]
      }
    },
    "/tenants/{tenantId}/civic/admin/memberships": {
      "get": {
        "summary": "List tenant memberships",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "memberships": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CivicTenantMembership"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Tenant administrator required"
          }
        },
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9][a-z0-9-]{1,62}$"
            }
          }
        ]
      }
    },
    "/tenants/{tenantId}/civic/admin/memberships/{userId}": {
      "put": {
        "summary": "Create or update tenant membership",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9][a-z0-9-]{1,62}$"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated"
          },
          "403": {
            "description": "Tenant administrator required"
          }
        }
      }
    },
    "/civic/admin/jurisdictions/{id}": {
      "put": {
        "summary": "Update a tenant jurisdiction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CivicJurisdictionMutation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated"
          },
          "400": {
            "description": "Invalid jurisdiction"
          },
          "403": {
            "description": "Tenant administrator required"
          },
          "404": {
            "description": "Jurisdiction not found"
          }
        }
      },
      "delete": {
        "summary": "Deactivate a tenant jurisdiction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deactivated"
          },
          "403": {
            "description": "Tenant administrator required"
          },
          "404": {
            "description": "Jurisdiction not found"
          },
          "409": {
            "description": "Active child jurisdictions must be deactivated first"
          }
        }
      }
    },
    "/tenants/{tenantId}/civic/admin/jurisdictions/{id}": {
      "put": {
        "summary": "Update a tenant jurisdiction",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CivicJurisdictionMutation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated"
          },
          "400": {
            "description": "Invalid jurisdiction"
          },
          "403": {
            "description": "Tenant administrator required"
          },
          "404": {
            "description": "Jurisdiction not found"
          }
        }
      },
      "delete": {
        "summary": "Deactivate a tenant jurisdiction",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deactivated"
          },
          "403": {
            "description": "Tenant administrator required"
          },
          "404": {
            "description": "Jurisdiction not found"
          },
          "409": {
            "description": "Active child jurisdictions must be deactivated first"
          }
        }
      }
    },
...
    "/civic/admin/membership-candidates": {
      "get": {
        "summary": "Search active users for tenant membership assignment",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 120
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CivicMembershipUser"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Search query is too short"
          },
          "403": {
            "description": "Tenant administrator required"
          }
        }
      }
    },
...
    "/tenants/{tenantId}/civic/admin/membership-candidates": {
      "get": {
        "summary": "Search active users for tenant membership assignment",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9][a-z0-9-]{1,62}$"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 120
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CivicMembershipUser"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Search query is too short"
          },
          "403": {
            "description": "Tenant administrator required"
          }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
...
      "CivicJurisdictionMutation": {
        "type": "object",
        "required": [
          "code",
          "levelKey",
          "name"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "levelKey": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "parentId": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
      "CivicMembershipUser": {
        "type": "object",
        "required": [
          "_id",
          "username"
        ],
        "properties": {
          "_id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "AgentBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "wt_agent_<clientId>.<secret>",
        "description": "Scoped API-client credential tied to an accountable user"
      }
    },
...
  },
...
}